Creamos la capa View de la arquitectura MVVM
- Movimos el MainActivity.kt al package view.
