--- app/src/main/java/com/example/mvvmapp/MainActivity.kt
@@ -1,15 +1,0 @@
-package com.example.mvvmapp
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.mvvmapp.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding:ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-}
