+++ app/src/main/java/com/example/mvvmapp/view/MainActivity.kt
@@ -1,0 +1,27 @@
+package com.example.mvvmapp.view
+
+import android.os.Bundle
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
+import com.example.mvvmapp.databinding.ActivityMainBinding
+import com.example.mvvmapp.viewmodel.QuoteViewModel
+
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding:ActivityMainBinding
+
+    private val quoteViewModel: QuoteViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        quoteViewModel.quoteModel.observe(this, Observer{ currentQuote ->
+            binding.tvQuote.text = currentQuote.quote
+            binding.tvAuthor.text = currentQuote.author
+        })
+
+        binding.viewContainer.setOnClickListener { quoteViewModel.randomQuote() }
+    }
+}
