+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:background="@color/app_background"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
     <TextView
         android:id="@+id/tvQuote"
